gem selling

+++ include/game/save_data.hpp
@@ -5,12 +5,12 @@
 
 class save_data final {
   private:
-    static uint64_t _pyrite_count;
-    static uint64_t _wolframite_count;
-    static uint64_t _carneol_count;
-    static uint64_t _moldavite_count;
-    static uint64_t _ruby_count;
-    static uint64_t _sapphire_count;
+    static int16_t _pyrite_count;
+    static int16_t _wolframite_count;
+    static int16_t _carneol_count;
+    static int16_t _moldavite_count;
+    static int16_t _ruby_count;
+    static int16_t _sapphire_count;
 
     static uint8_t _player_ship_speed_tier;
     static uint8_t _player_ship_damage_tier;
@@ -25,23 +25,23 @@ class save_data final {
     save_data() = default;
 
   public:
-    static uint64_t pyrite_count();
-    static void set_pyrite_count(uint64_t pyrite_count);
+    static int16_t pyrite_count();
+    static void set_pyrite_count(int16_t pyrite_count);
 
-    static uint64_t wolframite_count();
-    static void set_wolframite_count(uint64_t wolframite_count);
+    static int16_t wolframite_count();
+    static void set_wolframite_count(int16_t wolframite_count);
 
-    static uint64_t carneol_count();
-    static void set_carneol_count(uint64_t carneol_count);
+    static int16_t carneol_count();
+    static void set_carneol_count(int16_t carneol_count);
 
-    static uint64_t moldavite_count();
-    static void set_moldavite_count(uint64_t moldavite_count);
+    static int16_t moldavite_count();
+    static void set_moldavite_count(int16_t moldavite_count);
 
-    static uint64_t ruby_count();
-    static void set_ruby_count(uint64_t ruby_count);
+    static int16_t ruby_count();
+    static void set_ruby_count(int16_t ruby_count);
 
-    static uint64_t sapphire_count();
-    static void set_sapphire_count(uint64_t sapphire_count);
+    static int16_t sapphire_count();
+    static void set_sapphire_count(int16_t sapphire_count);
 
     static uint8_t player_ship_speed_tier();
     static void set_player_ship_speed_tier(uint8_t player_ship_speed_tier);
